pyproject.toml
[tool.poetry]
name = "fortitude"
version = "0.1.0"
description = ""
authors = ["Gabriel Couture <gacou54@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.1"
Django = "^3.1.5"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
parameterized = "^0.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
